refactored mongoconnection to use readDB and the DB files to access db links and names
