+++ src/main/Java/data_access/readDB/DB_files/AdminsCollectionName
@@ -1,0 +1,1 @@
+admins
