+++ src/main/Java/data_access/readDB/DB_files/DB_uri
@@ -1,0 +1,1 @@
+mongodb+srv://ibrahimbilal:Ibrahim123@cluster0.ngph3.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
