+++ src/main/Java/data_access/readDB/DB_files/EventPostersCollectionName
@@ -1,0 +1,1 @@
+eventPosters
