+++ src/main/Java/data_access/readDB/DB_files/EventsCollectionName
@@ -1,0 +1,1 @@
+events
